requirements.txt
streamlit
opencv-python-headless
numpy
pillow
scikit-learn
streamlit-cropper
